#7: Moved certain methods from Testutils to PathInfoCardViewModel
